WIP: Change the build system to enable building with Qt 6
This was built with all deprecated KF code disabled:
-DQT_MAJOR_VERSION=6 \
-DEXCLUDE_DEPRECATED_BEFORE_AND_AT=5.90.0 \
-DKF_DISABLE_DEPRECATED_BEFORE_AND_AT=0x055a00

+++ src/i18n/ktranscript.cpp
@@ -813,7 +813,7 @@ QJSValue Scriptface::fallback()
 
 QJSValue Scriptface::nsubs()
 {
-    return QJSValue(subList->size());
+    return QJSValue(static_cast<int>(subList->size()));
 }
 
 QJSValue Scriptface::subs(const QJSValue &index)
